+++ CHANGELOG.md
@@ -47,7 +47,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 
 ### Fixed
 
-- `RDF.DateTime`: store microseconds
+- `RDF.DateTime` and `RDF.Time` store microseconds
 - `RDF.DateTime`: '24:00:00' is a valid time in a xsd:dateTime; the dateTime 
   value so represented is the first instant of the following day
 - `RDF.LangString`: non-strings or the empty string as language produce invalid